# Speed up, fix fish_status_to_signal