--- a/share/functions/fish_status_to_signal.fish
+++ b/share/functions/fish_status_to_signal.fish
@@ -1,15 +1,10 @@
 function fish_status_to_signal --description "Print signal name from argument (\$status), or just argument"
+    __fish_make_completion_signals # Make sure signals are cached
     for arg in $argv
         if test $arg -gt 128
-            set -l sigix (math $arg-128)
-            if test $sigix -le 31 # 31 = number of named signals below
-                set -l signals HUP INT QUIT ILL TRAP ABRT BUS FPE KILL USR1 \
-                    SEGV USR2 PIPE ALRM TERM STKFLT CHLD CONT STOP TSTP \
-                    TTIN TTOU URG XCPU XFSZ VTALRM PROF WINCH IO PWR SYS
-                echo SIG$signals[$sigix]
-            else
-                echo $arg
-            end
+            # Important: each element of $__kill_signals is a string like "10 USR1"
+            string replace -r --filter "^$(math $arg-128) " SIG $__kill_signals |
+                read -l signal && echo $signal || echo $arg
         else
             echo $arg
         end