--- a/share/functions/fish_status_to_signal.fish
+++ b/share/functions/fish_status_to_signal.fish
@@ -3,8 +3,12 @@ function fish_status_to_signal --description "Print signal name from argument (\
     for arg in $argv
         if test $arg -gt 128
             # Important: each element of $__kill_signals is a string like "10 USR1"
-            string replace -r --filter '^'(math $arg-128)' ' SIG $__kill_signals |
-                read -l signal && echo $signal || echo $arg
+            if set -l signal_names (string replace -r --filter '^'(math $arg-128)' ' SIG $__kill_signals)
+                # Some signals have multiple mnemonics. Pick the first one.
+                echo $signal_names[1]
+            else
+                echo $arg
+            end
         else
             echo $arg
         end