## Description

Some super small changes to `fish_status_to_signal` that grant a massive speed boost.
Running `(count $signals)` is slow, and silly considering that we know the number already. 
And this list of signals is never, ever going to change, so it won't waste any developer time in the future.

```fish
# In config.fish
set -g list (seq 255)

function new_status_to_signal
    for arg in $argv
        if test $arg -gt 128
            set -l sigix (math $arg - 128)
            if test $sigix -le 31
                set -l signals SIGHUP SIGINT SIGQUIT SIGILL SIGTRAP SIGABRT SIGBUS \
                    SIGFPE SIGKILL SIGUSR1 SIGSEGV SIGUSR2 SIGPIPE SIGALRM \
                    SIGTERM SIGSTKFLT SIGCHLD SIGCONT SIGSTOP SIGTSTP \
                    SIGTTIN SIGTTOU SIGURG SIGXCPU SIGXFSZ SIGVTALRM \
                    SIGPROF SIGWINCH SIGIO SIGPWR SIGSYS
                echo $signals[$sigix]
            else
                echo $arg
            end
        else
            echo $arg
        end
    end
    return 0
end
```

```fish
ilan@arch ~> hyperfine --shell=fish "fish_status_to_signal \$list" "new_status_to_signal \$list"
Benchmark 1: fish_status_to_signal $list
  Time (mean ± σ):      23.4 ms ±   3.0 ms    [User: 15.6 ms, System: 10.0 ms]
  Range (min … max):    12.9 ms …  37.4 ms    103 runs
 
Benchmark 2: new_status_to_signal $list
  Time (mean ± σ):      11.7 ms ±   0.9 ms    [User: 7.3 ms, System: 6.0 ms]
  Range (min … max):     6.7 ms …  14.5 ms    177 runs
 
Summary
  'new_status_to_signal $list' ran
    1.99 ± 0.30 times faster than 'fish_status_to_signal $list'
```