@floam 2021-12-05
Yikes! This speedup is nice but this function is no bueno! For starters, SIGBUS (10, not 7) comes after SIGKILL and before SIGSEGV on my system.

7 is SIGEMT.