@floam 2021-12-05
Would you like to also fix that while you are improving this function? We actually cache these - see the completions for `kill`/`killall`. What you want to do is check for `$__kill_signals` otherwise run `__fish_make_completion_signals`.